Extract animation logic to useHintAnimation
